Fri, Mar 3, 2023, 3:36 PM +08:00

+++ readability/readability.c
@@ -40,8 +40,8 @@ int main(void)
 
  */
     //Compute Coleman-Liau index
-    float l = (letters / words) * 100;
-    float s = (sentences / words) * 100;
+    float l = (float) (letters / words) * 100;
+    float s = (float) (sentences / words) * 100;
     float index = (float) 0.0588 * l - 0.296 * s - 15.8;
 
     if (index < 1)
